Removed non private methods from ADAMAlignmentRecordContext.
