+++ adam-apis/src/main/scala/org/bdgenomics/adam/apis/java/JavaAlignmentRecordRDD.scala
@@ -19,7 +19,6 @@ package org.bdgenomics.adam.apis.java
 
 import org.apache.spark.api.java.JavaRDD
 import org.bdgenomics.adam.rdd.ADAMContext._
-import org.bdgenomics.adam.rdd.read.AlignmentRecordContext._
 import org.bdgenomics.formats.avro._
 import parquet.hadoop.metadata.CompressionCodecName
