+++ adam-cli/src/main/scala/org/bdgenomics/adam/cli/PrintTags.scala
@@ -23,7 +23,6 @@ import org.apache.spark.rdd.RDD
 import org.bdgenomics.adam.projections.AlignmentRecordField._
 import org.bdgenomics.adam.projections.Projection
 import org.bdgenomics.adam.rdd.ADAMContext._
-import org.bdgenomics.adam.rdd.read.AlignmentRecordContext._
 import org.bdgenomics.formats.avro.AlignmentRecord
 import org.kohsuke.args4j.{ Argument, Option }
