+++ adam-cli/src/main/scala/org/bdgenomics/adam/cli/Reads2Ref.scala
@@ -22,7 +22,6 @@ import org.apache.spark.SparkContext
 import org.apache.spark.rdd.RDD
 import org.bdgenomics.adam.predicates.UniqueMappedReadPredicate
 import org.bdgenomics.adam.rdd.ADAMContext._
-import org.bdgenomics.adam.rdd.read.AlignmentRecordContext._
 import org.bdgenomics.formats.avro.{ AlignmentRecord, Pileup }
 import org.kohsuke.args4j.{ Option => option, Argument }
